migrated provider package code from LcData to LcRest.ServiceProfessionalService
