+++ web/App_Code/LcData.cs
@@ -732,198 +732,6 @@ public static partial class LcData
         }
     }
     #endregion
-    #region Package Type (Provider Packages)
-    public class ProviderPackagesView
-    {
-        public dynamic Packages;
-        public dynamic PackagesDetails;
-        public Dictionary<int, dynamic> PackagesByID;
-        public Dictionary<int, List<dynamic>> PackagesDetailsByPackage;
-    }
-    public static dynamic GetProviderPackage(int providerPackageID)
-    {
-        using (var db = Database.Open("sqlloco"))
-        {
-            return db.QuerySingle(SQLSelectFromPackage + @"
-                WHERE   p.ProviderPackageID = @0
-            ", providerPackageID);
-        }
-    }
-    public static dynamic GetProviderPackageServiceAttributes(int providerPackageID)
-    {
-        using (var db = Database.Open("sqlloco"))
-        {
-            return db.Query(SQLGetPackageServiceAttributesByPackageID, providerPackageID, GetCurrentLanguageID(), GetCurrentCountryID());
-        }
-    }
-    public const string SQLSelectFromPackage = @"
-                SELECT  p.ProviderPackageID
-                        ,p.PricingTypeID
-                        ,p.ProviderUserID
-                        ,p.PositionID
-                        ,p.ProviderPackageName As Name
-                        ,p.ProviderPackageDescription As Description
-                        ,p.ProviderPackagePrice As Price
-                        ,p.ProviderPackageServiceDuration As ServiceDuration
-                        ,p.FirstTimeClientsOnly
-                        ,p.NumberOfSessions
-                        ,p.PriceRate
-                        ,p.PriceRateUnit
-                        ,p.IsPhone
-                        ,p.LanguageID
-                        ,p.CountryID
-                        ,p.CreatedDate
-                        ,p.UpdatedDate
-                        ,p.Active
-                        ,p.VisibleToClientID
-                FROM    ProviderPackage As P
-    ";
-    public const string SQLGetPackageServiceAttributesByPackageID = @"
-                SELECT  PD.ServiceAttributeID
-                        ,A.Name
-                        ,A.ServiceAttributeDescription
-                        ,P.ProviderPackageID
-                FROM    ProviderPackageDetail As PD
-                         INNER JOIN
-                        ProviderPackage As P
-                          ON P.ProviderPackageID = PD.ProviderPackageID
-                         INNER JOIN
-                        ServiceAttribute As A
-                          ON A.ServiceAttributeID = PD.ServiceAttributeID
-                            AND A.LanguageID = P.LanguageID AND A.CountryID = P.CountryID
-                WHERE   A.LanguageID = @1
-                        AND A.CountryID = @2
-                        AND PD.ProviderPackageID = @0
-                ORDER BY A.Name ASC
-    ";
-    public const string SQLGetPackagesByMulti = SQLSelectFromPackage + @"
-                         INNER JOIN
-                        PricingType As PT
-                          ON P.PricingTypeID = PT.PricingTypeID
-                            AND P.LanguageID = PT.LanguageID
-                            AND P.CountryID = PT.CountryID
-                         INNER JOIN
-                        PositionPricingType AS PPT
-                          ON PPT.PositionID = P.PositionID
-                            AND PPT.PricingTypeID = PT.PricingTypeID
-                            AND PPT.LanguageID = PT.LanguageID
-                            AND PPT.CountryID = PT.CountryID
-                            AND PPT.Active = 1
-                WHERE   p.ProviderUserID = @0
-                         AND (@1 = -1 OR P.PositionID = @1)
-                         AND 
-                        p.LanguageID = @2 AND p.CountryID = @3
-                         AND 
-                        p.Active = 1
-                         AND (@4 = -1 OR p.ProviderPackageID = @4)
-                         AND (@5 = -1 OR p.PricingTypeID = @5)
-                         AND (@6 = -1 OR P.IsAddOn = @6)
-                         AND (@7 = -999 OR P.VisibleToClientID = @7)
-                ORDER BY PT.DisplayRank ASC
-    ";
-    public const string SQLGetPackageServiceAttributesByMulti = @"
-                SELECT  PD.ServiceAttributeID
-                        ,A.Name
-                        ,A.ServiceAttributeDescription
-                        ,P.ProviderPackageID
-                FROM    ProviderPackageDetail As PD
-                         INNER JOIN
-                        ProviderPackage As P
-                          ON P.ProviderPackageID = PD.ProviderPackageID
-                         INNER JOIN
-                        ServiceAttribute As A
-                          ON A.ServiceAttributeID = PD.ServiceAttributeID
-                            AND A.LanguageID = P.LanguageID AND A.CountryID = P.CountryID
-                WHERE   P.ProviderUserID = @0
-                         AND (@1 = -1 OR P.PositionID = @1)
-                         AND P.LanguageID = @2 AND P.CountryID = @3
-                         AND PD.Active = 1 AND P.Active = 1
-                         AND (@4 = -1 OR P.ProviderPackageID = @4)
-                         AND (@5 = -1 OR P.PricingTypeID = @5)
-                         AND (@6 = -1 OR P.IsAddOn = @6)
-                ORDER BY A.Name ASC
-    ";
-
-    private static dynamic QueryPackagesByMulti(Database db, int providerUserID, int positionID = -1, int packageID = -1, int pricingTypeID = -1, bool? isAddon = null, int? visibleToClientID = null)
-    {
-        return db.Query(SQLGetPackagesByMulti,
-            providerUserID,
-            positionID,
-            GetCurrentLanguageID(),
-            GetCurrentCountryID(),
-            packageID,
-            pricingTypeID,
-            (isAddon.HasValue ? (isAddon.Value ? 1 : 0) : -1),
-            (visibleToClientID.HasValue ? visibleToClientID : -999)
-        );
-    }
-
-    private static dynamic QueryPackageServiceAttributesByMulti(Database db, int providerUserID, int positionID = -1, int packageID = -1, int pricingTypeID = -1, bool? isAddon = null)
-    {
-        return db.Query(SQLGetPackageServiceAttributesByMulti,
-            providerUserID,
-            positionID,
-            GetCurrentLanguageID(),
-            GetCurrentCountryID(),
-            packageID,
-            pricingTypeID,
-            (isAddon.HasValue ? (isAddon.Value ? 1 : 0) : -1)
-        );
-    }
-
-    private static ProviderPackagesView ProviderPackagesViewFromDB(dynamic packages, dynamic details)
-    {
-        // Create index of packages, Key:ID, Value:Package record
-        var index = new Dictionary<int, dynamic>(packages.Count);
-        foreach (var pak in packages)
-        {
-            index.Add(pak.ProviderPackageID, pak);
-        }
-        // Create index of packages details per package, Key:PackageID, Value:List of details records
-        var detailsIndex = new Dictionary<int, List<dynamic>>();
-        foreach (var det in details)
-        {
-            List<dynamic> detI = null;
-            if (detailsIndex.ContainsKey(det.ProviderPackageID))
-                detI = detailsIndex[det.ProviderPackageID];
-            else {
-                detI = new List<dynamic>();
-                detailsIndex.Add(det.ProviderPackageID, detI);
-            }
-            detI.Add(det);
-        }
-
-        return new ProviderPackagesView { Packages = packages, PackagesDetails = details, PackagesByID = index, PackagesDetailsByPackage = detailsIndex };
-    }
-
-    public static ProviderPackagesView GetPricingPackagesForClient(int providerUserID, int clientID)
-    {
-        dynamic packages, details;
-
-        using (var db = Database.Open("sqlloco"))
-        {
-            // Get the provider packages
-            packages = QueryPackagesByMulti(db, providerUserID, visibleToClientID: clientID);
-
-            details = QueryPackageServiceAttributesByMulti(db, providerUserID);
-        }
-
-        return ProviderPackagesViewFromDB(packages, details);
-    }
-
-    public static ProviderPackagesView GetPricingPackagesByProviderPosition(int providerUserID, int positionID, int packageID = -1, int pricingTypeID = -1, bool? isAddon = null)
-    {
-        dynamic packages, details;
-        using (var db = Database.Open("sqlloco")){
-            // Get the Provider Packages
-            packages = QueryPackagesByMulti(db, providerUserID, positionID, packageID, pricingTypeID, isAddon);
-
-            details = QueryPackageServiceAttributesByMulti(db, providerUserID, positionID, packageID, pricingTypeID, isAddon);
-        }
-
-        return ProviderPackagesViewFromDB(packages, details);
-    }
-    #endregion
     #endregion
 
     #region Alerts
